put mkdir into the backend
The notion of a "directory" may not be valid with a non-POSIX namespace (e.g.,
an object store), so put the mkdir down a level. Ideally, creating directories
should be relegated to the *_init functions so that the logic is not happening
at the "top-level" (PrependDir), but that's a little too much code churn for
now, so going with the easier option. Backends without a notion of directories
should simply spit out a warning or error.